+++ src/structures/paging/mapper/mapped_page_table/mod.rs
@@ -1,3 +1,11 @@
+mod display;
+mod iter;
+mod offset_page_table;
+mod range_iter;
+
+pub use self::display::Display;
+#[cfg(target_pointer_width = "64")]
+pub use self::offset_page_table::{OffsetPageTable, PhysOffset};
 use crate::structures::paging::{
     mapper::*,
     page::AddressNotAligned,
@@ -82,7 +90,7 @@ impl<P: PageTableFrameMapping> Mapper<Size1GiB> for MappedPageTable<'_, P> {
     fn unmap(
         &mut self,
         page: Page<Size1GiB>,
-    ) -> Result<(PhysFrame<Size1GiB>, MapperFlush<Size1GiB>), UnmapError> {
+    ) -> Result<(PhysFrame<Size1GiB>, PageTableFlags, MapperFlush<Size1GiB>), UnmapError> {
         let p4 = &mut self.level_4_table;
         let p3 = self
             .page_table_walker
@@ -102,7 +110,39 @@ impl<P: PageTableFrameMapping> Mapper<Size1GiB> for MappedPageTable<'_, P> {
             .map_err(|AddressNotAligned| UnmapError::InvalidFrameAddress(p3_entry.addr()))?;
 
         p3_entry.set_unused();
-        Ok((frame, MapperFlush::new(page)))
+        Ok((frame, flags, MapperFlush::new(page)))
+    }
+
+    fn clear(&mut self, page: Page<Size1GiB>) -> Result<UnmappedFrame<Size1GiB>, UnmapError> {
+        let p4 = &mut self.level_4_table;
+        let p3 = self
+            .page_table_walker
+            .next_table_mut(&mut p4[page.p4_index()])?;
+
+        let p3_entry = &mut p3[page.p3_index()];
+        let flags = p3_entry.flags();
+
+        if !flags.contains(PageTableFlags::HUGE_PAGE) {
+            return Err(UnmapError::ParentEntryHugePage);
+        }
+
+        if !flags.contains(PageTableFlags::PRESENT) {
+            let cloned = p3_entry.clone();
+            p3_entry.set_unused();
+            return Ok(UnmappedFrame::NotPresent { entry: cloned });
+        }
+
+        let frame = PhysFrame::from_start_address(p3_entry.addr())
+            .map_err(|AddressNotAligned| UnmapError::InvalidFrameAddress(p3_entry.addr()))?;
+        let flags = p3_entry.flags();
+
+        p3_entry.set_unused();
+
+        Ok(UnmappedFrame::Present {
+            frame,
+            flags,
+            flush: MapperFlush::new(page),
+        })
     }
 
     unsafe fn update_flags(
@@ -207,7 +247,7 @@ impl<P: PageTableFrameMapping> Mapper<Size2MiB> for MappedPageTable<'_, P> {
     fn unmap(
         &mut self,
         page: Page<Size2MiB>,
-    ) -> Result<(PhysFrame<Size2MiB>, MapperFlush<Size2MiB>), UnmapError> {
+    ) -> Result<(PhysFrame<Size2MiB>, PageTableFlags, MapperFlush<Size2MiB>), UnmapError> {
         let p4 = &mut self.level_4_table;
         let p3 = self
             .page_table_walker
@@ -230,7 +270,40 @@ impl<P: PageTableFrameMapping> Mapper<Size2MiB> for MappedPageTable<'_, P> {
             .map_err(|AddressNotAligned| UnmapError::InvalidFrameAddress(p2_entry.addr()))?;
 
         p2_entry.set_unused();
-        Ok((frame, MapperFlush::new(page)))
+        Ok((frame, flags, MapperFlush::new(page)))
+    }
+
+    fn clear(&mut self, page: Page<Size2MiB>) -> Result<UnmappedFrame<Size2MiB>, UnmapError> {
+        let p4 = &mut self.level_4_table;
+        let p3 = self
+            .page_table_walker
+            .next_table_mut(&mut p4[page.p4_index()])?;
+        let p2 = self
+            .page_table_walker
+            .next_table_mut(&mut p3[page.p3_index()])?;
+
+        let p2_entry = &mut p2[page.p2_index()];
+        let flags = p2_entry.flags();
+
+        if !flags.contains(PageTableFlags::HUGE_PAGE) {
+            return Err(UnmapError::ParentEntryHugePage);
+        }
+
+        if !flags.contains(PageTableFlags::PRESENT) {
+            let cloned = p2_entry.clone();
+            p2_entry.set_unused();
+            return Ok(UnmappedFrame::NotPresent { entry: cloned });
+        }
+        let frame = PhysFrame::from_start_address(p2_entry.addr())
+            .map_err(|AddressNotAligned| UnmapError::InvalidFrameAddress(p2_entry.addr()))?;
+        let flags = p2_entry.flags();
+
+        p2_entry.set_unused();
+        Ok(UnmappedFrame::Present {
+            frame,
+            flags,
+            flush: MapperFlush::new(page),
+        })
     }
 
     unsafe fn update_flags(
@@ -357,7 +430,7 @@ impl<P: PageTableFrameMapping> Mapper<Size4KiB> for MappedPageTable<'_, P> {
     fn unmap(
         &mut self,
         page: Page<Size4KiB>,
-    ) -> Result<(PhysFrame<Size4KiB>, MapperFlush<Size4KiB>), UnmapError> {
+    ) -> Result<(PhysFrame<Size4KiB>, PageTableFlags, MapperFlush<Size4KiB>), UnmapError> {
         let p4 = &mut self.level_4_table;
         let p3 = self
             .page_table_walker
@@ -371,13 +444,47 @@ impl<P: PageTableFrameMapping> Mapper<Size4KiB> for MappedPageTable<'_, P> {
 
         let p1_entry = &mut p1[page.p1_index()];
 
-        let frame = p1_entry.frame().map_err(|err| match err {
+        let frame = p1_entry.frame(true).map_err(|err| match err {
             FrameError::FrameNotPresent => UnmapError::PageNotMapped,
-            FrameError::HugeFrame => UnmapError::ParentEntryHugePage,
+            FrameError::HugeFrame => unreachable!(),
         })?;
+        let flags = p1_entry.flags();
 
         p1_entry.set_unused();
-        Ok((frame, MapperFlush::new(page)))
+        Ok((frame, flags, MapperFlush::new(page)))
+    }
+
+    fn clear(&mut self, page: Page<Size4KiB>) -> Result<UnmappedFrame<Size4KiB>, UnmapError> {
+        let p4 = &mut self.level_4_table;
+        let p3 = self
+            .page_table_walker
+            .next_table_mut(&mut p4[page.p4_index()])?;
+        let p2 = self
+            .page_table_walker
+            .next_table_mut(&mut p3[page.p3_index()])?;
+        let p1 = self
+            .page_table_walker
+            .next_table_mut(&mut p2[page.p2_index()])?;
+
+        let p1_entry = &mut p1[page.p1_index()];
+
+        let frame = match p1_entry.frame(true) {
+            Ok(frame) => frame,
+            Err(FrameError::HugeFrame) => unreachable!(),
+            Err(FrameError::FrameNotPresent) => {
+                let cloned = p1_entry.clone();
+                p1_entry.set_unused();
+                return Ok(UnmappedFrame::NotPresent { entry: cloned });
+            }
+        };
+        let flags = p1_entry.flags();
+
+        p1_entry.set_unused();
+        Ok(UnmappedFrame::Present {
+            frame,
+            flags,
+            flush: MapperFlush::new(page),
+        })
     }
 
     unsafe fn update_flags(
@@ -616,7 +723,7 @@ impl<P: PageTableFrameMapping> CleanUp for MappedPageTable<'_, P> {
                                 Page::range_inclusive(start, end),
                                 frame_deallocator,
                             ) {
-                                let frame = entry.frame().unwrap();
+                                let frame = entry.frame(false).unwrap();
                                 entry.set_unused();
                                 frame_deallocator.deallocate_frame(frame);
                             }
@@ -665,7 +772,7 @@ impl<P: PageTableFrameMapping> PageTableWalker<P> {
     ) -> Result<&'b PageTable, PageTableWalkError> {
         let page_table_ptr = self
             .page_table_frame_mapping
-            .frame_to_pointer(entry.frame()?);
+            .frame_to_pointer(entry.frame(false)?);
         let page_table: &PageTable = unsafe { &*page_table_ptr };
 
         Ok(page_table)
@@ -683,7 +790,7 @@ impl<P: PageTableFrameMapping> PageTableWalker<P> {
     ) -> Result<&'b mut PageTable, PageTableWalkError> {
         let page_table_ptr = self
             .page_table_frame_mapping
-            .frame_to_pointer(entry.frame()?);
+            .frame_to_pointer(entry.frame(false)?);
         let page_table: &mut PageTable = unsafe { &mut *page_table_ptr };
 
         Ok(page_table)